Add MFA device helper functions.
