+++ api/types/authentication.go
@@ -17,7 +17,6 @@ limitations under the License.
 package types
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -27,7 +26,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/gogo/protobuf/jsonpb"
 	"github.com/gravitational/trace"
 
 	"github.com/gravitational/teleport/api/constants"
@@ -939,107 +937,6 @@ func (wal *WebauthnLocalAuth) Check() error {
 	return nil
 }
 
-// NewMFADevice creates a new MFADevice with the given name. Caller must set
-// the Device field in the returned MFADevice.
-func NewMFADevice(name, id string, addedAt time.Time) *MFADevice {
-	return &MFADevice{
-		Metadata: Metadata{
-			Name: name,
-		},
-		Id:       id,
-		AddedAt:  addedAt,
-		LastUsed: addedAt,
-	}
-}
-
-// setStaticFields sets static resource header and metadata fields.
-func (d *MFADevice) setStaticFields() {
-	d.Kind = KindMFADevice
-	d.Version = V1
-}
-
-// CheckAndSetDefaults validates MFADevice fields and populates empty fields
-// with default values.
-func (d *MFADevice) CheckAndSetDefaults() error {
-	d.setStaticFields()
-	if err := d.Metadata.CheckAndSetDefaults(); err != nil {
-		return trace.Wrap(err)
-	}
-	if d.Id == "" {
-		return trace.BadParameter("MFADevice missing ID field")
-	}
-	if d.AddedAt.IsZero() {
-		return trace.BadParameter("MFADevice missing AddedAt field")
-	}
-	if d.LastUsed.IsZero() {
-		return trace.BadParameter("MFADevice missing LastUsed field")
-	}
-	if d.LastUsed.Before(d.AddedAt) {
-		return trace.BadParameter("MFADevice LastUsed field must be earlier than AddedAt")
-	}
-	if d.Device == nil {
-		return trace.BadParameter("MFADevice missing Device field")
-	}
-	if err := checkWebauthnDevice(d); err != nil {
-		return trace.Wrap(err)
-	}
-	return nil
-}
-
-func checkWebauthnDevice(d *MFADevice) error {
-	wrapper, ok := d.Device.(*MFADevice_Webauthn)
-	if !ok {
-		return nil
-	}
-	switch webDev := wrapper.Webauthn; {
-	case webDev == nil:
-		return trace.BadParameter("MFADevice has malformed WebauthnDevice")
-	case len(webDev.CredentialId) == 0:
-		return trace.BadParameter("WebauthnDevice missing CredentialId field")
-	case len(webDev.PublicKeyCbor) == 0:
-		return trace.BadParameter("WebauthnDevice missing PublicKeyCbor field")
-	default:
-		return nil
-	}
-}
-
-func (d *MFADevice) GetKind() string         { return d.Kind }
-func (d *MFADevice) GetSubKind() string      { return d.SubKind }
-func (d *MFADevice) SetSubKind(sk string)    { d.SubKind = sk }
-func (d *MFADevice) GetVersion() string      { return d.Version }
-func (d *MFADevice) GetMetadata() Metadata   { return d.Metadata }
-func (d *MFADevice) GetName() string         { return d.Metadata.GetName() }
-func (d *MFADevice) SetName(n string)        { d.Metadata.SetName(n) }
-func (d *MFADevice) GetRevision() string     { return d.Metadata.GetRevision() }
-func (d *MFADevice) SetRevision(rev string)  { d.Metadata.SetRevision(rev) }
-func (d *MFADevice) Expiry() time.Time       { return d.Metadata.Expiry() }
-func (d *MFADevice) SetExpiry(exp time.Time) { d.Metadata.SetExpiry(exp) }
-
-// MFAType returns the human-readable name of the MFA protocol of this device.
-func (d *MFADevice) MFAType() string {
-	switch d.Device.(type) {
-	case *MFADevice_Totp:
-		return "TOTP"
-	case *MFADevice_U2F:
-		return "U2F"
-	case *MFADevice_Webauthn:
-		return "WebAuthn"
-	default:
-		return "unknown"
-	}
-}
-
-func (d *MFADevice) MarshalJSON() ([]byte, error) {
-	buf := new(bytes.Buffer)
-	err := (&jsonpb.Marshaler{}).Marshal(buf, d)
-	return buf.Bytes(), trace.Wrap(err)
-}
-
-func (d *MFADevice) UnmarshalJSON(buf []byte) error {
-	unmarshaler := jsonpb.Unmarshaler{AllowUnknownFields: true}
-	err := unmarshaler.Unmarshal(bytes.NewReader(buf), d)
-	return trace.Wrap(err)
-}
 
 // IsSessionMFARequired returns whether this RequireMFAType requires per-session MFA.
 func (r RequireMFAType) IsSessionMFARequired() bool {
